Finished moving to separate routers
