+++ server/routes/api.router.js
@@ -1,6 +1,6 @@
 const express = require("express");
 
-const userRouter = require('./user');
+const {userRouter} = require('./user');
 const {adminRouter} = require('./admin');
 const {orderRouter} = require('./index');
